refactor: CheckEventTarget.js에서 productPurchaseMenu 관련 파일 분리
- CheckProductPurchaseMenu.js로 파일 분리해서 관리
